requirements.txt
Flask
Flask-SocketIO
google-cloud-speech
python-dotenv
